refactor: move assignment endpoints to corresponding module
