+++ backend/src/assignments/assignments.module.ts
@@ -1,9 +1,11 @@
+import { UsersModule } from '@/users/users.module';
 import { Module } from '@nestjs/common';
 
 import { AssignmentsController } from './assignments.controller';
 import { AssignmentsService } from './assignments.service';
 
 @Module({
+  imports: [UsersModule],
   controllers: [AssignmentsController],
   providers: [AssignmentsService],
 })
