+++ backend/src/assignments/dto/auto-assign.dto.ts.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty, IsUUID } from 'class-validator';
+
+export class AutoAssignDto {
+  @IsNotEmpty()
+  @IsUUID()
+  courseId!: string;
+}
